refactor: remove magic number
